+++ src/ace_skyspark_lib/__init__.py
@@ -9,7 +9,7 @@ from ace_skyspark_lib.models.history import (
     TimeRange,
 )
 
-__version__ = "0.1.26"
+__version__ = "0.1.27"
 
 __all__ = [
     "Equipment",